Manual lint fixes in packages/core
