+++ packages/core/examples/contextMenuExample.tsx
@@ -3,6 +3,8 @@
  * Licensed under the Apache License, Version 2.0 - http://www.apache.org/licenses/LICENSE-2.0
  */
 
+// tslint:disable max-classes-per-file
+
 import * as classNames from "classnames";
 import * as PureRender from "pure-render-decorator";
 import * as React from "react";
