+++ packages/core/src/components/forms/controls.tsx
@@ -3,6 +3,9 @@
  * Licensed under the Apache License, Version 2.0 - http://www.apache.org/licenses/LICENSE-2.0
  */
 
+// HACKHACK: these components should go in separate files
+// tslint:disable max-classes-per-file
+
 import * as classNames from "classnames";
 import * as React from "react";
