+++ packages/core/src/components/forms/inputGroup.tsx
@@ -66,7 +66,7 @@ export class InputGroup extends React.Component<HTMLInputProps & IInputGroupProp
 
         return (
             <div className={classes}>
-                { leftIconName == null ? null : <span className={`pt-icon ${Classes.iconClass(leftIconName)}`} /> }
+                {leftIconName == null ? null : <span className={`pt-icon ${Classes.iconClass(leftIconName)}`} />}
 
                 <input
                     type="text"
