+++ packages/core/src/components/hotkeys/hotkeysTarget.tsx
@@ -23,7 +23,6 @@ export interface IHotkeysTarget extends React.Component<any, any>, React.Compone
     renderHotkeys(): React.ReactElement<IHotkeysProps>;
 }
 
-/* tslint:disable:no-invalid-this */
 export function HotkeysTarget<T extends { prototype: IHotkeysTarget }>(constructor: T) {
     const { componentWillMount, componentWillUnmount, render, renderHotkeys } = constructor.prototype;
 
@@ -31,6 +30,8 @@ export function HotkeysTarget<T extends { prototype: IHotkeysTarget }>(construct
         throw new Error(`@HotkeysTarget-decorated class must implement \`renderHotkeys\`. ${constructor}`);
     }
 
+    // tslint:disable:no-invalid-this
+    // tslint:disable-next-line only-arrow-functions
     constructor.prototype.componentWillMount = function() {
         this.localHotkeysEvents = new HotkeysEvents(HotkeyScope.LOCAL);
         this.globalHotkeysEvents = new HotkeysEvents(HotkeyScope.GLOBAL);
@@ -44,6 +45,7 @@ export function HotkeysTarget<T extends { prototype: IHotkeysTarget }>(construct
         }
     };
 
+    // tslint:disable-next-line only-arrow-functions
     constructor.prototype.componentWillUnmount = function() {
         // detach global key event listeners
         document.removeEventListener("keydown", this.globalHotkeysEvents.handleKeyDown);
@@ -57,6 +59,7 @@ export function HotkeysTarget<T extends { prototype: IHotkeysTarget }>(construct
         }
     };
 
+    // tslint:disable-next-line only-arrow-functions
     constructor.prototype.render = function() {
         const element = render.call(this) as JSX.Element;
 
@@ -84,5 +87,5 @@ export function HotkeysTarget<T extends { prototype: IHotkeysTarget }>(construct
             return element;
         }
     };
+    // tslint:enable:no-invalid-this
 };
-/* tslint:enable:no-invalid-this */
