+++ packages/core/test/tabs/tabsTests.tsx
@@ -3,6 +3,8 @@
  * Licensed under the Apache License, Version 2.0 - http://www.apache.org/licenses/LICENSE-2.0
  */
 
+// tslint:disable max-classes-per-file
+
 import { assert } from "chai";
 import { mount, ReactWrapper } from "enzyme";
 import * as React from "react";
